chore(api): refactor reset password
