+++ packages/api/src/resolvers/reset_password_verify.ts
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken'
 import { DecodedToken } from '../types'
-import { findUser } from '../queries'
+import * as UsersAPI from '../graphql/users'
 
 export default async (root, args): Promise<DecodedToken> => {
   const { token } = args
@@ -12,7 +12,7 @@ export default async (root, args): Promise<DecodedToken> => {
   const now = String(new Date().getTime()).substring(0, 10)
   if (Number(now) > Number(expired_at)) throw new Error('invalid_token')
 
-  const user = await findUser({ id: decoded.id, reset_password_token: token })
+  const user = (await UsersAPI.find({ id: decoded.id, reset_password_token: token }))[0]
   if (!user) throw new Error('invalid_token')
 
   return decoded
